feat: removes authconfig creation
- dashboard related traffic is now protected by using [single AuthConfig](https://github.com/opendatahub-io/opendatahub-operator/pull/605/commits/7924666795c139dc707ef5e2bb22b5b0d4179efc) in the AppNamespace
- removed logic. tests and assets related to AuthConfig creation
