+++ Makefile
@@ -21,7 +21,6 @@ help: ## Display this help.
 .PHONY: generate
 generate: tools ## Generates required resources for the controller to work properly (see config/ folder)
 	$(LOCALBIN)/controller-gen rbac:roleName=manager-role crd webhook paths="./..." output:crd:artifacts:config=config/crd/bases
-	$(call fetch-external-crds,github.com/kuadrant/authorino,api/v1beta1)
 	$(call fetch-external-crds,github.com/openshift/api,route/v1)
 
 SRC_DIRS:=./controllers ./test
