+++ controllers/mesh_env_config.go
@@ -2,16 +2,11 @@ package controllers
 
 import (
 	"os"
-	"strings"
-
-	"github.com/pkg/errors"
 )
 
 const (
-	MeshNamespaceEnv       = "MESH_NAMESPACE"
-	ControlPlaneEnv        = "CONTROL_PLANE_NAME"
-	AuthorinoLabelSelector = "AUTHORINO_LABEL"
-	AuthAudience           = "AUTH_AUDIENCE"
+	MeshNamespaceEnv = "MESH_NAMESPACE"
+	ControlPlaneEnv  = "CONTROL_PLANE_NAME"
 )
 
 func getControlPlaneName() string {
@@ -20,28 +15,6 @@ func getControlPlaneName() string {
 
 func getMeshNamespace() string {
 	return getEnvOr(MeshNamespaceEnv, "istio-system")
-}
-
-func getAuthorinoLabel() ([]string, error) {
-	label := getEnvOr(AuthorinoLabelSelector, "authorino/topic=odh")
-	keyValue := strings.Split(label, "=")
-
-	if len(keyValue) != 2 {
-		return nil, errors.Errorf("Expected authorino label to be in key=value format, got [%s]", label)
-	}
-
-	return keyValue, nil
-}
-
-func getAuthAudience() []string {
-	aud := getEnvOr(AuthAudience, "https://kubernetes.default.svc")
-	audiences := strings.Split(aud, ",")
-
-	for i := range audiences {
-		audiences[i] = strings.TrimSpace(audiences[i])
-	}
-
-	return audiences
 }
 
 func getEnvOr(key, defaultValue string) string {
