--- controllers/template/authconfig.yaml
@@ -1,33 +1,0 @@
-apiVersion: authorino.kuadrant.io/v1beta1
-kind: AuthConfig
-metadata:
-  name: odh-dashboard-protection
-  namespace: $ODH_NS
-  labels:
-    authorino/topic: odh
-spec:
-  hosts:
-    - "${ODH_ROUTE}"
-  identity:
-    - name: kubernetes-users
-      kubernetes:
-        audiences:
-          - "https://kubernetes.default.svc"
-  authorization:
-    - name: k8s-rbac
-      kubernetes:
-        user:
-          valueFrom: { authJSON: auth.identity.username }
-  response:
-    - name: x-auth-data
-      json:
-        properties:
-          - name: username
-            valueFrom: { authJSON: auth.identity.username }
-  denyWith:
-    unauthenticated:
-      message:
-        value: "Access denied"
-    unauthorized:
-      message:
-        value: "Unauthorized"
