--- test/test_assets.go
@@ -1,6 +1,0 @@
-package test
-
-import _ "embed"
-
-//go:embed data/expected_authconfig.yaml
-var ExpectedAuthConfig []byte
